+++ src/core/CMakeLists.txt
@@ -32,6 +32,7 @@ set(SRCS
             frontend/camera/factory.cpp
             frontend/camera/interface.cpp
             frontend/emu_window.cpp
+            frontend/framebuffer_layout.cpp
             frontend/motion_emu.cpp
             gdbstub/gdbstub.cpp
             hle/config_mem.cpp
@@ -216,6 +217,7 @@ set(HEADERS
             frontend/camera/factory.h
             frontend/camera/interface.h
             frontend/emu_window.h
+            frontend/framebuffer_layout.h
             frontend/input.h
             frontend/motion_emu.h
             gdbstub/gdbstub.h
