+++ src/core/frontend/framebuffer_layout.h
